- Add foldrAsFoldl implementation - Adapt FStream to work on binary trees (add new Steps and unfold and fold operations) - Add classes representing optimization steps for unfoldr/foldl, foldrrecursive and foldrAsFoldl - Add Fibonacci example (unfold/fold on binary trees).

+++ src/optimizations/optimizations_unfoldr_foldl/NoFunctions.java
@@ -27,7 +27,7 @@ public class NoFunctions {
                 over = true;
             } else {
                 auxState = aux.get().getY();
-                value = ((BiFunction<Integer, Integer, Integer>) (x, y) -> x * y).apply(value, (Integer) aux.get().getX());
+                value = ((BiFunction<Integer, Integer, Integer>) (x, y) -> x * y).apply(value, aux.get().getX());
             }
         }
